Initiate unit test setup
Unit tests are done using python3's stdlib module `unittest`.
See the README.md on how to run the unit tests.
This needed some restructuring, but this way we got also rid of the "GH_WORKSPACE" env var. The action now assumes that its current working directory is the root of the mod repository.