Don't fail whole validation for invalid remote files
I reconsidered the handling of remote files.
While they technically have to be valid according to the spec/schema from AVC,
it does not really make sense to fail the whole validation when there's an error.

So this commit reduces it from an error to a warning.

There are a lot of cases where this is unwanted:
* Creating a new version file, where the remote file isn't uploaded yet
    (likely because the file you are creating is actually the one that will be uploaded/pushed later)
* Fixing a broken (remote) version file:
    you are in the process of fixing the problem, and you want to know whether your fixes actually work.
    Failing the PR validation due to an error you have probably fixed right now is not practical.
* So basically all cases where the remote file and the "local" file are the same.

Only for cases where the remote version file is a separate file, maybe hosted elsewhere,
it makes sense to fail hard. But these are probably the minority.
There will be an option added in the future to make the validation fail hard in case of invalid remote files,
for those users where it makes sense.